pyproject.toml
[project]
name = "dataproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "duckdb>=1.4.1",
    "fastapi[standard]>=0.123.9",
    "pendulum>=3.1.0",
    "polars>=1.34.0",
    "polygon-api-client>=1.16.2",
    "pyarrow>=22.0.0",
    "requests>=2.32.5",
]

[tool.ruff]
line-length = 80
